Finish up discovery only
